style: standardize catalog view css
