+++ src/components/catalog/views/navigation/CatalogBreadcrumbView.tsx
@@ -10,17 +10,17 @@ export const CatalogBreadcrumbView: FC<{}> = () =>
     if(!activeNodes || activeNodes.length === 0)
     {
         return (
-            <div className="nitro-catalog-classic-breadcrumb">
+            <div className="nitro-catalog-breadcrumb">
                 <span>{ LocalizeText('catalog.title') }</span>
             </div>
         );
     }
 
     return (
-        <div className="nitro-catalog-classic-breadcrumb">
+        <div className="nitro-catalog-breadcrumb">
             { activeNodes.map((node, index) => (
-                <span key={ node.pageId } className="nitro-catalog-classic-breadcrumb-segment">
-                    <span className="nitro-catalog-classic-breadcrumb-separator">&rsaquo;</span>
+                <span key={ node.pageId } className="nitro-catalog-breadcrumb-segment">
+                    <span className="nitro-catalog-breadcrumb-separator">&rsaquo;</span>
                     <span
                         className={ `truncate ${ index === activeNodes.length - 1 ? 'font-semibold' : 'cursor-pointer hover:underline' }` }
                         onClick={ index < activeNodes.length - 1 ? () => activateNode(node) : undefined }
