+++ src/components/vault/VaultView.tsx
@@ -66,22 +66,6 @@ const rewardCurrencyType = (reward: IEarningsReward): number | string | null =>
     }
 };
 
-// Scoped colour override for the Guadagni window only: classic blue header +
-// cool grey body (the shared 'primary-slim' theme is teal + cream). Higher
-// specificity (.nitro-card.nitro-vault ...) than the theme so it wins. The body
-// element renders `.nitro-card-content-shell`, NOT `.content-area`.
-const VAULT_STYLES = `
-  .nitro-card.nitro-vault .nitro-card-header {
-    background: linear-gradient(180deg, #5a80b8 0%, #3f63a0 100%);
-    border-color: #34548a;
-  }
-  .nitro-card.nitro-vault,
-  .nitro-card.nitro-vault .content-area,
-  .nitro-card.nitro-vault .nitro-card-content-shell {
-    background: #dde1e6 !important;
-  }
-`;
-
 export const VaultView: FC<{}> = props =>
 {
     const [ isVisible, setIsVisible ] = useState(false);
@@ -184,9 +168,7 @@ export const VaultView: FC<{}> = props =>
     return (
         <NitroCardView className="nitro-vault w-[430px]" theme="primary-slim" uniqueKey="vault">
             <NitroCardHeaderView headerText={ localizeWithFallback('earnings.title', 'Guadagni') } onCloseClick={ () => setIsVisible(false) } />
-            <NitroCardContentView className="flex flex-col gap-[3px] text-black">
-                <style>{ VAULT_STYLES }</style>
-                { CATEGORIES.map(category =>
+            <NitroCardContentView className="flex flex-col gap-[3px] text-black">                { CATEGORIES.map(category =>
                 {
                     const entry = entriesByKey.get(category.key) ?? null;
                     const canClaim = !!entry && entry.enabled && entry.claimable;
