+++ src/index.tsx
@@ -19,7 +19,7 @@ import './css/index.css';
 
 import './css/backgrounds/BackgroundsView.css';
 import './css/badges/BadgeLeaderboardView.css';
-import './css/catalog/CatalogClassicView.css';
+import './css/catalog/CatalogView.css';
 import './css/emustats/EmuStatsView.css';
 
 import './css/chat/Chats.css';
